requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
chardet==5.2.0
chumpy==0.70
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
contourpy==1.0.5
cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
embreex==2.17.7.post3
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fqdn==1.5.1
h5py==3.9.0
icecream==2.1.3
importlib-metadata==6.8.0
importlib_resources==6.0.1
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-loader==0.3
matplotlib-inline==0.1.6
mesh-to-depth==0.1.2
mistune==3.0.2
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.23.1
numpy-stl==3.0.1
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycollada==0.7.2
pygments==2.16.1
pyhocon==0.3.57
pymcubes==0.1.2
pyntcloud==0.3.1
pysdf==0.1.9
python-json-logger==2.0.7
python-utils==3.8.1
qtconsole==5.4.4
qtpy==2.4.1
referencing==0.30.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rtree==1.0.1
scikit-image==0.21.0
send2trash==1.8.2
shapely==2.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
svg-path==6.3
terminado==0.17.1
tifffile==2023.9.18
tinycss2==1.2.1
tomli==2.0.1
torchgeometry==0.1.2
torchviz==0.0.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
trimesh==3.23.5
types-python-dateutil==2.8.19.14
uri-template==1.3.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xatlas==0.0.8
xxhash==3.3.0
pyyaml

